requirements.txt
fastapi
httpx
litellm
yt-dlp
youtube-transcript-api
numpy<2
numba==0.62.1
llvmlite==0.45.1
pyacoustid
openai-whisper
trafilatura
beautifulsoup4
pypdf
pdfplumber
python-docx
python-pptx
openpyxl
striprtf
Pillow
pillow-avif-plugin
pytesseract
python-dotenv
python-multipart
playwright
essentia
torch==2.2.2
transformers==4.48.3
timm
uvicorn
